Complete nep141 cli commands. Split into separate files
